+++ partial/examples/examples.js
@@ -57,6 +57,16 @@ angular.module('generic.table').controller('DocumentationController',function($s
             }
         ],
         data:[{
+            "name":"gt-id (optional)",
+            "type":"number",
+            "default":"id of table scope",
+            "description":"unique id for table, if none is passed, the scope id will be returned. Use id together with events"
+        },{
+            "name":"gt-classes (optional)",
+            "type":"string",
+            "default":"",
+            "description":"add classes to table element e.g. table-bordered, table-condensed etc. for bootstrap styles"
+        },{
             "name":"gt-data (optional)",
             "type":"array",
             "default":"",
@@ -193,21 +203,21 @@ angular.module('generic.table').controller('DocumentationController',function($s
             }
         ],
         data:[{
-            "name":"gt-update-table",
+            "name":"gt-update-table:gtId",
             "description":"update table data",
-            "options":"data array, ie. $scope.$broadcast('gt-update-table', data);"
+            "options":"data array, ie. $scope.$broadcast('gt-update-table:tableId', data);"
         },{
-            "name":"gt-update-structure",
+            "name":"gt-update-structure:gtId",
             "description":"update structure of table (settings and field definitions)",
-            "options":"object, ie. $scope.$broadcast('gt-update-structure', {fields:[],settings:[]});"
+            "options":"object, ie. $scope.$broadcast('gt-update-structure:tableId', {fields:[],settings:[]});"
         },{
-            "name":"gt-paginate-table",
+            "name":"gt-paginate-table:gtId",
             "description":"change how many rows are visible",
-            "options":"number, ie. $scope.$broadcast('gt-paginate-table', 10);"
+            "options":"number, ie. $scope.$broadcast('gt-paginate-table:tableId', 10);"
         },{
-            "name":"gt-export-csv",
+            "name":"gt-export-csv:gtId",
             "description":"export table data to csv, <a href='/#examples#exportOptions'>see all export options</a>.",
-            "options":"export settings object, ie. $scope.$broadcast('gt-export-csv', {fileName:'my-custom-export'});"
+            "options":"export settings object, ie. $scope.$broadcast('gt-export-csv:tableId', {fileName:'my-custom-export'});"
         }]
     };
 
@@ -346,10 +356,10 @@ angular.module('generic.table').controller('DocumentationController',function($s
         }]
     };
 
-}).controller('CustomRenderExampleController',function($scope, mockService,$filter,$sce){
+}).controller('CustomRenderExampleController',function($scope, mockService){
 
     $scope.exportCsv = function() {
-        $scope.$broadcast('gt-export-csv');
+        $scope.$broadcast('gt-export-csv:'+$scope.tableCustomRender.id);
     };
 
 
@@ -384,6 +394,14 @@ angular.module('generic.table').controller('DocumentationController',function($s
                 sort:'enable',
                 sortOrder:0,
                 columnOrder:2
+            },{
+                objectKey:'details',
+                visible:true,
+                enabled:true,
+                sort:'disable',
+                sortOrder:0,
+                columnOrder:5,
+                export:false
             }
 
         ],
@@ -393,7 +411,7 @@ angular.module('generic.table').controller('DocumentationController',function($s
                 type:"STRING",
                 objectKey:'favoriteColor',
                 classNames:"text-right middle",
-                render:function(row){return $sce.trustAsHtml('<div style="float:right;width:15px;height:15px;border-radius:50%;background: '+row.favoriteColor+'"></div>')}
+                render:function(row){return '<div style="float:right;width:15px;height:15px;border-radius:50%;background: '+row.favoriteColor+'"></div>'}
             },{
                 name:"Birthday",
                 type:"DATE",
@@ -412,26 +430,34 @@ angular.module('generic.table').controller('DocumentationController',function($s
                 classNames:"text-right",
                 render: function(row){return moment().diff(moment.unix(row.birthday),'years')},
                 value: function(row){return moment().diff(moment.unix(row.birthday),'years')}
+            },{
+                name:'Details',
+                type:"NUMBER",
+                objectKey:'details',
+                classNames:"text-right",
+                render: function(row, column){ return '<a>{{row.isOpen ? "Hide":"Show"}}<a/>'},
+                compile:true,
+                value: function(row){return 'details'},
+                expand:'<custom-dir></custom-dir>'
             }
         ],
         data:[]
     };
 
     mockService.getJsonData().then(function(res){
-        $scope.$broadcast("gt-update-table", res);
+        $scope.$broadcast("gt-update-table:"+$scope.tableCustomRender.id, res);
     });
 
 
 }).factory('mockService',function($resource, $http) {
 
     /* remote service example
-    function getData () {
-        var url = 'http://example.com/json.json';
-        return $http.jsonp(url).then(function (response) {
-            return response.data;
-        });
-
-    }*/
+     function getData () {
+     var url = 'http://example.com/json.json';
+     return $http.jsonp(url).then(function (response) {
+     return response.data;
+     });
+     }*/
 
     function getJsonData(){
         var url = './partial/examples/mock-data.json';
@@ -443,5 +469,19 @@ angular.module('generic.table').controller('DocumentationController',function($s
 
     return {
         "getJsonData":getJsonData
+    };
+}).directive('customDir', function() {
+    return {
+        replace:true,
+        template:'<div><div class="pull-left"><label>Name:&nbsp;</label>{{row.fullName}}<br><label>Age:&nbsp;</label>{{row.age}}<br><label>Favorite random number:&nbsp;</label>{{randomNumber}}<br><a ng-click="update()">Update</a></div> <a class="pull-right" ng-click="close()">Close</a> </div>',
+        restrict: 'E',
+        link: function(scope, element, attrs, fn) {
+            scope.close = scope.toggleRow; // assign close function to button in directive
+            scope.randomNumber = Math.random();
+            scope.update = function(){
+                scope.randomNumber = Math.random();
+                console.log('update something');
+            };
+        }
     };
 });